refactor : Repository Class를 interface로 추상화 시킴
